+++ test/lib/android/src/main/java/com/adjust/test/AdjustSdkTest.java
@@ -71,4 +71,14 @@ public class AdjustSdkTest extends ReactContextBaseJavaModule {
     public void setTestConnectionOptions() {
         TestConnectionOptions.setTestConnectionOptions();
     }
+
+    @ReactMethod
+    public void addListener(String eventName) {
+        // Keep: Required for RN built in Event Emitter Calls.
+    }
+
+    @ReactMethod
+    public void removeListeners(double count) {
+        // Keep: Required for RN built in Event Emitter Calls.
+    }
 }